article/requirements.txt
<<<<<<< HEAD
asgiref==3.5.2
Django==4.0.6
django-froala-editor==4.0.13
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
=======
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-countries==7.6.1
django-extensions==3.2.3
django-froala-editor==4.3.0
djangorestframework==3.15.2
idna==3.10
pillow==10.4.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
>>>>>>> a8d18b23b80ec99c1d3244ee281b19dc054e7ea5
